refactor(stellar): modularize deploy-contract.js functionality (#644)
