+++ stellar/utils.js
@@ -12,10 +12,11 @@ const {
     xdr,
     nativeToScVal,
 } = require('@stellar/stellar-sdk');
-const { downloadContractCode, VERSION_REGEX, SHORT_COMMIT_HASH_REGEX } = require('../common/utils');
+const { downloadContractCode, VERSION_REGEX, SHORT_COMMIT_HASH_REGEX, isValidStellarAddress } = require('../common/utils');
 const { printInfo, sleep, addEnvOption, getCurrentVerifierSet } = require('../common');
 const { Option } = require('commander');
 const { ethers } = require('hardhat');
+const { itsCustomMigrationDataToScValV110 } = require('./type-utils');
 const {
     utils: { arrayify, hexZeroPad, id, isHexString, keccak256 },
     BigNumber,
@@ -37,6 +38,14 @@ const SUPPORTED_CONTRACTS = new Set([
     'Upgrader',
     'Multicall',
 ]);
+
+const CustomMigrationDataTypeToScValV110 = {
+    InterchainTokenService: (migrationData) => itsCustomMigrationDataToScValV110(migrationData),
+};
+
+const VERSIONED_CUSTOM_MIGRATION_DATA_TYPES = {
+    '1.1.0': CustomMigrationDataTypeToScValV110,
+};
 
 function getNetworkPassphrase(networkType) {
     switch (networkType) {
@@ -186,7 +195,7 @@ async function broadcast(operation, wallet, chain, action, options = {}, simulat
     }
 
     const tx = await prepareTransaction(operation, server, wallet, chain.networkType, options);
-    return await sendTransaction(tx, server, action, options);
+    return sendTransaction(tx, server, action, options);
 }
 
 function getAssetCode(balance, chain) {
@@ -295,7 +304,7 @@ const getNewSigners = async (wallet, config, chain, options) => {
         };
     }
 
-    return await getAmplifierVerifiers(config, chain.axelarId);
+    return getAmplifierVerifiers(config, chain.axelarId);
 };
 
 function serializeValue(value) {
@@ -417,7 +426,18 @@ const getContractCodePath = async (options, contractName) => {
 
     if (options.version) {
         const url = getContractR2Url(contractName, options.version);
-        return await downloadContractCode(url, contractName, options.version);
+        return downloadContractCode(url, contractName, options.version);
+    }
+
+    throw new Error('Either --artifact-path or --version must be provided');
+};
+
+const getUploadContractCodePath = async (options, contractName) => {
+    if (options.artifactPath) return options.artifactPath;
+
+    if (options.version) {
+        const url = getContractR2Url(contractName, options.version);
+        return downloadContractCode(url, contractName, options.version);
     }
 
     throw new Error('Either --artifact-path or --version must be provided');
@@ -463,6 +483,59 @@ function pascalToKebab(str) {
     return str.replace(/([A-Z])/g, (match, _, offset) => (offset > 0 ? `-${match.toLowerCase()}` : match.toLowerCase()));
 }
 
+function sanitizeMigrationData(migrationData, version, contractName) {
+    if (migrationData === null || migrationData === '()') return null;
+
+    if (isValidStellarAddress(migrationData)) {
+        return migrationData;
+    }
+
+    if (typeof migrationData === 'string') {
+        return migrationData;
+    }
+
+    let parsed;
+
+    try {
+        parsed = JSON.parse(migrationData);
+    } catch (_) {
+        // not json, continue
+    }
+
+    if (Array.isArray(parsed)) {
+        return parsed.map((value) => sanitizeMigrationData(value, version, contractName));
+    }
+
+    const custom = customMigrationData(parsed, version, contractName);
+
+    if (custom) {
+        return custom;
+    }
+
+    if (parsed !== null && typeof parsed === 'object') {
+        return Object.fromEntries(Object.entries(parsed).map(([key, value]) => [key, sanitizeMigrationData(value, version, contractName)]));
+    }
+
+    printInfo('Sanitized migration data', parsed);
+
+    return parsed;
+}
+
+function customMigrationData(migrationDataObj, version, contractName) {
+    if (!version || !VERSIONED_CUSTOM_MIGRATION_DATA_TYPES[version] || !VERSIONED_CUSTOM_MIGRATION_DATA_TYPES[version][contractName]) {
+        return null;
+    }
+
+    const customMigrationDataTypeToScVal = VERSIONED_CUSTOM_MIGRATION_DATA_TYPES[version][contractName];
+
+    try {
+        printInfo(`Retrieving custom migration data for ${contractName}`);
+        return customMigrationDataTypeToScVal(migrationDataObj);
+    } catch (error) {
+        throw new Error(`Failed to convert custom migration data for ${contractName}: ${error}`);
+    }
+}
+
 module.exports = {
     stellarCmd,
     ASSET_TYPE_NATIVE,
@@ -484,8 +557,10 @@ module.exports = {
     tokenMetadataToScVal,
     saltToBytes32,
     getContractCodePath,
+    getUploadContractCodePath,
     isValidAddress,
     SUPPORTED_CONTRACTS,
     BytesToScVal,
     pascalToKebab,
+    sanitizeMigrationData,
 };
