+++ sui/contract.js
@@ -150,17 +150,7 @@ async function pause(keypair, client, chain, args, options) {
     contract.disallowedFunctions.versions = contract.disallowedFunctions.versions.concat(versionsArg);
     contract.disallowedFunctions.functionNames = contract.disallowedFunctions.functionNames.concat(allowedFunctionsArg);
 
-    return await disallowFunctions(
-        keypair,
-        client,
-        packageId,
-        moduleName,
-        singletonId,
-        ownerCapId,
-        versionsArg,
-        allowedFunctionsArg,
-        options,
-    );
+    return disallowFunctions(keypair, client, packageId, moduleName, singletonId, ownerCapId, versionsArg, allowedFunctionsArg, options);
 }
 
 async function unpause(keypair, client, chain, args, options) {
@@ -199,7 +189,7 @@ async function unpause(keypair, client, chain, args, options) {
         }
     }
 
-    return await allowFunctions(keypair, client, packageId, moduleName, singletonId, ownerCapId, versionsArg, allowedFunctionsArg, options);
+    return allowFunctions(keypair, client, packageId, moduleName, singletonId, ownerCapId, versionsArg, allowedFunctionsArg, options);
 }
 
 async function processCommand(command, chain, args, options) {
